Add files via upload
This project was a collaborative effort between Comp Sci and CIT majors. The project goal was decided before i joined the team, the goal of the project was to create a new system for remote login to track students. This was done in response to the pandemic forcing classes online. The files here were my  contribution to the project